user login system complete with styling

+++ backend/app/controllers/delete_controller.rb
@@ -2,8 +2,13 @@ class DeleteController < ApplicationController
     skip_before_action :verify_authenticity_token
 
     def create
-        account = Account.find_by('email': params[:email])
-        profile = Profile.find_by('email': params[:email])
+        account = Account.find_by('email': session[:user_email])
+        profile = Profile.find_by('email': session[:user_email])
+
+        if (account.nil? && profile.nil?)
+            render json: { :success => false, :message => "You are not logged in." }
+            return
+        end
 
         account.destroy
         profile.destroy
